Move translation script out of pages to fix exportPathMap error
